// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/CustomerDocument.ts
export interface CustomerDocument {
    /**
     * 
     * @type {string}
     * @memberof CustomerDocument
     */
    documentId?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerDocument
     */
    documentName?: string;
    /**
     * 
     * @type {Date}
     * @memberof CustomerDocument
     */
    statementStartDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CustomerDocument
     */
    statementEndDay?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CustomerDocument
     */
    documentDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof CustomerDocument
     */
    documentType?: CustomerDocumentDocumentTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CustomerDocument
     */
    documentSource?: CustomerDocumentDocumentSourceEnum;
}


/**
 * @export
 */
export const CustomerDocumentDocumentTypeEnum = {
    LoanDocument: 'LOAN_DOCUMENT',
    Statement: 'STATEMENT'
} as const;
export type CustomerDocumentDocumentTypeEnum = typeof CustomerDocumentDocumentTypeEnum[keyof typeof CustomerDocumentDocumentTypeEnum];

/**
 * @export
 */
export const CustomerDocumentDocumentSourceEnum = {
    Imported: 'IMPORTED',
    Generated: 'GENERATED'
} as const;
export type CustomerDocumentDocumentSourceEnum = typeof CustomerDocumentDocumentSourceEnum[keyof typeof CustomerDocumentDocumentSourceEnum];